Update references after portal class renames
