+++ blockly/src/client/Client.java
@@ -30,8 +30,8 @@ import java.util.Properties;
 import java.util.Set;
 import java.util.function.Supplier;
 import level.produs.*;
-import portal.MyCalculations;
-import portal.MyEnergyPelletCatcherBehavior;
+import portal.BlocklyEnergyPelletCatcherBehavior;
+import portal.BlocklyPortalCalculations;
 import portal.PortalRegistry;
 import portal.controlls.Hero;
 import portal.portals.PortalColor;
@@ -155,8 +155,8 @@ public class Client {
           createHero();
           createSystems();
           PortalRegistry.setDebugMode(false);
-          PortalRegistry.registerCalculations(() -> new MyCalculations());
-          PortalRegistry.registerPelletCatcherBehavior(() -> new MyEnergyPelletCatcherBehavior());
+          PortalRegistry.registerCalculations(() -> new BlocklyPortalCalculations());
+          PortalRegistry.registerPelletCatcherBehavior(() -> new BlocklyEnergyPelletCatcherBehavior());
 
           startServer();
         });
